chore(client-side-test): implemet test for the client side
implement feedback from Lf
