+++ client/src/components/auth/Signup.jsx
@@ -3,29 +3,32 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
 // validations
-import signinValidator from '../validation/signinValidator';
+import signupValidator from '../../validation/signupValidator';
 
 // actions
-import fetchUserSignin from '../actions/userSignIn';
+import fetchUserSignup from '../../actions/signup';
 
 
 /**
  *
  *
- * @class Signin
+ * @class Signup
  * @extends {React.Component}
  */
-class Signin extends React.Component {
+class Signup extends React.Component {
   /**
-   * Creates an instance of Signin.
+   * Creates an instance of Signup.
    * @param {any} props
-   * @memberof Signin
+   * @memberof Signup
    */
   constructor(props) {
     super(props);
     this.state = {
+      fullname: '',
+      username: '',
       email: '',
       password: '',
+      confirmPassword: '',
       errors: {}
     };
     this.onChange = this.onChange.bind(this);
@@ -33,24 +36,24 @@ class Signin extends React.Component {
   }
 
   /**
-  *
-  * @returns {json} with nextProps details
-  * @param {any} nextProps
-  * @memberof Signin
-  */
+ *
+ * @returns {json} with nextProps details
+ * @param {any} nextProps
+ * @memberof Signup
+ */
   componentWillReceiveProps(nextProps) {
     if (nextProps.auth.isAuthenticated &&
       nextProps.auth.errorMessage.length === 0) {
-      $('#signin-modal').modal('hide');
+      $('#signup-modal').modal('hide');
     }
   }
 
   /**
- * @description handles form change events
- * @returns {null} null
- * @param {any} event
- * @memberof Signin
- */
+* @description handles form change events
+* @returns {null} null
+* @param {any} event
+* @memberof Signup
+*/
   onChange(event) {
     this.setState({ [event.target.name]: event.target.value });
   }
@@ -60,24 +63,24 @@ class Signin extends React.Component {
  *
  * @param {any} event
  * @returns {dispatch} react-redux dispatch
- * @memberof Signin
+ * @memberof Signup
  */
   onSubmit(event) {
     event.preventDefault();
     const isValid = this.isValid();
     if (isValid) {
-      this.props.signinUser(this.state);
+      this.props.signupUser(this.state);
     }
   }
 
   /**
-*
-*
-* @returns {boolean} boolean
-* @memberof Signin
-*/
+ *
+ *
+ * @returns {boolean} boolean
+ * @memberof Signup
+ */
   isValid() {
-    const { isValid, errors } = signinValidator(this.state);
+    const { isValid, errors } = signupValidator(this.state);
     if (!isValid) {
       this.setState({ errors });
     } else {
@@ -90,21 +93,23 @@ class Signin extends React.Component {
    * @description react render method
    *
    * @returns {component} react component
-   * @memberof Signin
+   * @memberof Signup
    */
   render() {
     const { errors } = this.state;
 
-    let signinError;
+    let signupError;
     if (this.props.auth.errorMessage) {
-      signinError = (
-        <span className="help-block text-danger">{this.props.auth.errorMessage}</span>
+      signupError = (
+        <span className="help-block text-danger">
+          {this.props.auth.errorMessage}
+        </span>
       );
     }
     return (
       <div>
         <div
-          id="signin-modal"
+          id="signup-modal"
           className="modal fade"
         >
           <div className="modal-dialog modal-dialog-centered modal-sm">
@@ -127,7 +132,7 @@ class Signin extends React.Component {
               </button>
               <div className="modal-body">
 
-                {signinError}
+                {signupError}
 
                 <div className="row">
                   <div className="col-sm-12">
@@ -135,12 +140,46 @@ class Signin extends React.Component {
 
                       <div className="form-group">
                         <input
+                          type="text"
+                          name="fullname"
+                          className="form-control"
+                          placeholder="Full Name"
+                          value={this.state.fullname}
+                          onChange={this.onChange}
+                        />
+                        {
+                          errors.fullname &&
+                          <span className="help-block text-danger">
+                            {errors.fullname}
+                          </span>
+                        }
+                      </div>
+
+                      <div className="form-group">
+                        <input
+                          type="text"
+                          name="username"
+                          className="form-control"
+                          placeholder="User Name"
+                          value={this.state.username}
+                          onChange={this.onChange}
+                        />
+                        {
+                          errors.username &&
+                          <span className="help-block text-danger">
+                            {errors.username}
+                          </span>
+                        }
+                      </div>
+
+                      <div className="form-group">
+                        <input
                           type="email"
-                          placeholder="Email"
                           name="email"
+                          className="form-control"
+                          placeholder="Email Address"
                           value={this.state.email}
                           onChange={this.onChange}
-                          className="form-control signin-input"
                         />
                         {
                           errors.email &&
@@ -153,16 +192,34 @@ class Signin extends React.Component {
                       <div className="form-group">
                         <input
                           type="password"
-                          placeholder="Password"
                           name="password"
+                          className="form-control"
+                          placeholder="Password"
                           value={this.state.password}
                           onChange={this.onChange}
-                          className="form-control signin-input"
                         />
                         {
                           errors.password &&
                           <span className="help-block text-danger">
                             {errors.password}
+                          </span>
+                        }
+                      </div>
+
+
+                      <div className="form-group">
+                        <input
+                          type="password"
+                          name="confirmPassword"
+                          className="form-control"
+                          placeholder="Confirm Password"
+                          value={this.state.confirmPassword}
+                          onChange={this.onChange}
+                        />
+                        {
+                          errors.confirmPassword &&
+                          <span className="help-block text-danger">
+                            {errors.confirmPassword}
                           </span>
                         }
                       </div>
@@ -183,7 +240,6 @@ class Signin extends React.Component {
     );
   }
 }
-
 const authProps = {
   isAuthenticated: PropTypes.bool,
   user: PropTypes.string,
@@ -191,14 +247,14 @@ const authProps = {
   errorMessage: PropTypes.string
 };
 
-Signin.propTypes = {
+Signup.propTypes = {
   auth: PropTypes.shape(authProps).isRequired,
-  signinUser: PropTypes.func.isRequired
+  signupUser: PropTypes.func.isRequired
 };
 
 const mapDispatchToProps = dispatch => ({
-  signinUser: (userDetails) => {
-    dispatch(fetchUserSignin(userDetails));
+  signupUser: (userDetails) => {
+    dispatch(fetchUserSignup(userDetails));
   }
 });
 
@@ -206,4 +262,4 @@ const mapStateToProps = state => ({
   auth: state.auth
 });
 
-export default connect(mapStateToProps, mapDispatchToProps)(Signin);
+export default connect(mapStateToProps, mapDispatchToProps)(Signup);
