+++ client/src/components/recipe/SearchRecipe.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import searchRecipes from '../actions/searchRecipe';
+import searchRecipes from '../../actions/searchRecipe';
 /**
  *
  *
