+++ client/src/css/style.css
@@ -309,3 +309,7 @@ margin-top: 150px;
   0% { transform: rotate(0deg); }
   100% { transform: rotate(360deg); }
 }
+
+.user-name {
+  cursor: pointer;
+}
