+++ server/test/api/recipe.spec.js
@@ -6,6 +6,9 @@ chai.use(chaiHttp);
 
 let userToken;
 let recipeId;
+const userToken2 =
+  'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MywiaWF0IjoxNTE5MTI3NDc3fQ.jOZ8BYDMtc0M6ajJougAkN3Uq_NTI1mq7wwVb1ZaL2o';
+
 
 describe('RECIPE CONTROLLER', () => {
   before((done) => {
@@ -38,6 +41,7 @@ describe('RECIPE CONTROLLER', () => {
         })
         .end((error, response) => {
           expect(response.status).to.equal(400);
+          expect(response.body).to.be.an('object');
           expect(response.body.error.name)
             .to.include('Recipe name is required');
           done();
@@ -56,6 +60,7 @@ describe('RECIPE CONTROLLER', () => {
         })
         .end((error, response) => {
           expect(response.status).to.equal(400);
+          expect(response.body).to.be.an('object');
           expect(response.body.error.description)
             .to.include('Description for recipe is required');
           done();
@@ -74,6 +79,7 @@ describe('RECIPE CONTROLLER', () => {
         })
         .end((error, response) => {
           expect(response.status).to.equal(400);
+          expect(response.body).to.be.an('object');
           expect(response.body.error.ingredients)
             .to.include('ingredients for recipe is required');
           done();
@@ -91,6 +97,20 @@ describe('RECIPE CONTROLLER', () => {
         })
         .end((error, response) => {
           expect(response.status).to.equal(201);
+          expect(response.body).to.be.an('object');
+          expect(response.body.message).to.equal('Recipe created successfully');
+          expect(response.body.recipe.id)
+            .to.be.a('number');
+          expect(response.body.recipe.name)
+            .to.be.a('string');
+          expect(response.body.recipe.userId)
+            .to.be.a('number');
+          expect(response.body.recipe.ingredients)
+            .to.be.a('string');
+          expect(response.body.recipe.description)
+            .to.be.a('string');
+          expect(response.body.recipe.recipeImage)
+            .to.be.a('string');
           done();
         });
     });
@@ -109,11 +129,18 @@ describe('RECIPE CONTROLLER', () => {
         .end((error, response) => {
           recipeId = response.body.recipe.id;
           expect(response.status).to.equal(201);
+          expect(response.body).to.be.an('object');
           expect(response.body.recipe.name).to.equal('Test Recipe');
           expect(response.body.recipe.ingredients)
             .to.equal('test, test, test');
           expect(response.body.recipe.description)
             .to.equal('For testing the recipe');
+          expect(response.body.recipe.recipeImage)
+            .to.be.a('string');
+          expect(response.body.recipe.id)
+            .to.be.a('number');
+          expect(response.body.recipe.userId)
+            .to.be.a('number');
           done();
         });
     });
@@ -129,6 +156,7 @@ describe('RECIPE CONTROLLER', () => {
         })
         .end((error, response) => {
           expect(response.status).to.equal(401);
+          expect(response.body).to.be.an('object');
           expect(response.body.message).to.equal('Unauthorized.');
           done();
         });
@@ -148,6 +176,8 @@ describe('RECIPE CONTROLLER', () => {
         })
         .end((error, response) => {
           expect(response.status).to.equal(404);
+          expect(response.body).to.be.an('object');
+          expect(response.body.message).to.be.a('string');
           done();
         });
     });
@@ -165,6 +195,7 @@ describe('RECIPE CONTROLLER', () => {
           })
           .end((error, response) => {
             expect(response.status).to.equal(401);
+            expect(response.body).to.be.an('object');
             expect(response.body.message).to.equal('Unauthorized.');
             done();
           });
@@ -183,11 +214,18 @@ describe('RECIPE CONTROLLER', () => {
         })
         .end((error, response) => {
           expect(response.status).to.equal(200);
+          expect(response.body).to.be.an('object');
           expect(response.body.recipe.name).to.equal('Another Test Recipe');
           expect(response.body.recipe.ingredients)
             .to.equal('testing, testing, testing');
           expect(response.body.recipe.description)
             .to.equal('For testing the recipe again');
+          expect(response.body.recipe.recipeImage)
+            .to.be.a('string');
+          expect(response.body.recipe.id)
+            .to.be.a('number');
+          expect(response.body.recipe.userId)
+            .to.be.a('number');
           done();
         });
     });
@@ -206,7 +244,9 @@ describe('RECIPE CONTROLLER', () => {
           })
           .end((error, response) => {
             expect(response).to.have.status(400);
-            expect(response.body.message).to.equal('RecipeId must be a number');
+            expect(response.body).to.be.an('object');
+            expect(response.body.error.recipeId)
+              .to.include('RecipeId must be a number');
             done();
           });
       }
@@ -220,6 +260,18 @@ describe('RECIPE CONTROLLER', () => {
         .get('/api/v1/recipes')
         .end((error, response) => {
           expect(response).to.have.status(200);
+          expect(response.body).to.be.an('object');
+          expect(response.body.numberOfItems)
+            .to.be.a('number');
+          expect(response.body.limit)
+            .to.be.a('number');
+          expect(response.body.currentPage)
+            .to.be.a('number');
+          expect(response.body.pages)
+            .to.be.a('number');
+          expect(response.body.recipes)
+            .to.be.an('array').with.lengthOf(3);
+          expect(response.body).to.have.property('recipes').with.lengthOf(3);
           done();
         });
     });
@@ -231,6 +283,10 @@ describe('RECIPE CONTROLLER', () => {
           .get(`/api/v1/recipes/${recipeId}`)
           .end((error, response) => {
             expect(response).to.have.status(200);
+            expect(response.body).to.be.an('object');
+            expect(response.body.recipe)
+              .to.be.an('object');
+            expect(response.body).to.have.property('recipe');
             done();
           });
       }
@@ -242,9 +298,37 @@ describe('RECIPE CONTROLLER', () => {
         .set('token', userToken)
         .end((error, response) => {
           expect(response).to.have.status(200);
+          expect(response.body).to.be.an('object');
+          expect(response.body.numberOfItems)
+            .to.be.a('number');
+          expect(response.body.limit)
+            .to.be.a('number');
+          expect(response.body.currentPage)
+            .to.be.a('number');
+          expect(response.body.pages)
+            .to.be.a('number');
+          expect(response.body.recipes)
+            .to.be.an('array').with.lengthOf(2);
+          expect(response.body).to.have.property('recipes').with.lengthOf(2);
           done();
         });
     });
+
+
+    it(
+      'Should not allow a user that does not exist view all their recipes',
+      (done) => {
+        chai.request(app)
+          .get('/api/v1/recipes/user/recipes')
+          .set('token', userToken2)
+          .end((error, response) => {
+            expect(response).to.have.status(404);
+            expect(response.body.message)
+              .to.equal('A user with that id is not found');
+            done();
+          });
+      }
+    );
 
 
     it('Should not allow non auth user view all their recipes', (done) => {
@@ -252,6 +336,7 @@ describe('RECIPE CONTROLLER', () => {
         .get('/api/v1/recipes/user/recipes')
         .end((error, response) => {
           expect(response).to.have.status(401);
+          expect(response.body).to.be.an('object');
           expect(response.body.message).to.equal('Unauthorized.');
           done();
         });
@@ -264,7 +349,9 @@ describe('RECIPE CONTROLLER', () => {
           .get('/api/v1/recipes/recipeId')
           .end((error, response) => {
             expect(response).to.have.status(400);
-            expect(response.body.message).to.equal('RecipeId must be a number');
+            expect(response.body).to.be.an('object');
+            expect(response.body.error.recipeId)
+              .to.include('RecipeId must be a number');
             done();
           });
       }
@@ -278,6 +365,8 @@ describe('RECIPE CONTROLLER', () => {
         .set('token', userToken)
         .end((error, response) => {
           expect(response.status).to.equal(404);
+          expect(response.body).to.be.an('object');
+          expect(response.body.message).to.be.a('string');
           done();
         });
     });
@@ -287,6 +376,7 @@ describe('RECIPE CONTROLLER', () => {
         .delete(`/api/v1/recipes/${recipeId}`)
         .end((error, response) => {
           expect(response.status).to.equal(401);
+          expect(response.body).to.be.an('object');
           expect(response.body.message).to.equal('Unauthorized.');
           done();
         });
@@ -298,6 +388,8 @@ describe('RECIPE CONTROLLER', () => {
         .set('token', userToken)
         .end((error, response) => {
           expect(response.status).to.equal(200);
+          expect(response.body).to.be.an('object');
+          expect(response.body.message).to.equal('recipe deleted');
           done();
         });
     });
@@ -309,10 +401,48 @@ describe('RECIPE CONTROLLER', () => {
           .get('/api/v1/recipes/recipeId')
           .end((error, response) => {
             expect(response).to.have.status(400);
-            expect(response.body.message).to.equal('RecipeId must be a number');
+            expect(response.body).to.be.an('object');
+            expect(response.body.error.recipeId)
+              .to.include('RecipeId must be a number');
             done();
           });
       }
     );
+  });
+
+  describe('Search for a recipe', () => {
+    it('Should return 404 for a recipe that does not exist', (done) => {
+      chai.request(app)
+        .post('/api/v1/recipes/search?search=eba')
+        .end((error, response) => {
+          expect(response).to.have.status(404);
+          expect(response.body).to.be.an('object');
+          expect(response.body.message).to.be.a('string');
+          expect(response.body.message).to.equal('No match(es) found');
+          done();
+        });
+    });
+
+    it('Should return 200 for a recipe that exist', (done) => {
+      chai.request(app)
+        .post('/api/v1/recipes/search?search=test')
+        .end((error, response) => {
+          expect(response).to.have.status(200);
+          expect(response.body).to.be.an('object');
+          expect(response.body.numberOfItems)
+            .to.be.a('number');
+          expect(response.body.limit)
+            .to.be.a('number');
+          expect(response.body.currentPage)
+            .to.be.a('number');
+          expect(response.body.pages)
+            .to.be.a('number');
+          expect(response.body.recipes)
+            .to.be.an('array').with.lengthOf(2);
+          expect(response.body)
+            .to.have.property('recipes').with.lengthOf(2);
+          done();
+        });
+    });
   });
 });
